refactor(my-orders): move resolveOrderStatusLabel to _utils
src/app/AGENTS.md separates _utils (route-local pure functions) from
_constants (route-local constants), but labels.ts held a branching pure
function together with its two lookup tables.

The TDD gate exempts **/_constants/** on the premise that the folder holds
no logic, so a function needing verification sat in an exempt path. Move
resolveOrderStatusLabel and its private tables to _utils/orderStatusLabel.ts
with their four test cases; ORDER_STATUS_TAB_LABELS and PAY_METHOD_LABEL
stay behind as plain constants.

Delete labels.component.test.ts along with it. Its remaining case only
compared constant literals, and every other path in the gate's exclude list
carries no test file — this was the sole exception. Behavior is unchanged.

Closes #224
