+++ src/app/(main)/(my-order)/my-orders/_components/OrderCard.tsx
@@ -18,7 +18,8 @@ import type { OrderListItem, OrderStatus } from "@/core/domain/order";
 import { CUSTOMER_INPUT_ROUTES } from "@/core/domain/product-category";
 import { routes } from "@/core/domain/routes";
 import { cancelOrder } from "@/actions/cancelOrder";
-import { PAY_METHOD_LABEL, resolveOrderStatusLabel } from "@/app/(main)/(my-order)/my-orders/_constants/labels";
+import { PAY_METHOD_LABEL } from "@/app/(main)/(my-order)/my-orders/_constants/labels";
+import { resolveOrderStatusLabel } from "@/app/(main)/(my-order)/my-orders/_utils/orderStatusLabel";
 import { getInvitationInputDaysLeft } from "@/app/(main)/(my-order)/my-orders/_utils/orderDeadline";
 import { PaymentButton } from "./PaymentButton";
 import { PendingCoupleInfoBanner } from "./PendingCoupleInfoBanner";
