+++ src/app/(main)/(my-order)/my-orders/_utils/orderStatusLabel.component.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { ORDER_STATUS_TAB_LABELS, resolveOrderStatusLabel } from "./labels";
+import { resolveOrderStatusLabel } from "./orderStatusLabel";
 import { MOBILE_INVITATION_CATEGORY } from "@/core/domain/product-category";
 
 describe("resolveOrderStatusLabel", () => {
@@ -22,10 +22,5 @@ describe("resolveOrderStatusLabel", () => {
   it("카테고리별 예외가 없는 상태는 기본 라벨을 그대로 쓴다", () => {
     expect(resolveOrderStatusLabel("PENDING", MOBILE_INVITATION_CATEGORY)).toBe("주문대기");
     expect(resolveOrderStatusLabel("CANCELLED", MOBILE_INVITATION_CATEGORY)).toBe("취소");
-  });
-
-  it("탭 라벨은 카테고리가 섞이는 자리라 중립어를 쓴다", () => {
-    expect(ORDER_STATUS_TAB_LABELS.COMPLETED).toBe("완료");
-    expect(ORDER_STATUS_TAB_LABELS.CONFIRMED).toBe("결제완료");
   });
 });
